+++ recipe-guru.Model/Requests/RatingSearchRequest.cs
@@ -3,7 +3,7 @@ namespace recipe_guru.Model.Requests
 {
     public class RatingSearchRequest
     {
-        public long ReceptId { get; set; }
+        public int ReceptId { get; set; }
 
         public RatingMark Mark { get; set; }
     }
