+++ recipe-guru.Model/Requests/ReceptKoraciSearchRequest.cs
@@ -3,6 +3,6 @@ namespace recipe_guru.Model.Requests
 {
     public class ReceptKoraciSearchRequest
     {
-        public long ReceptId { get; set; }
+        public int ReceptId { get; set; }
     }
 }
